+++ R/mx_api.R
@@ -17,8 +17,9 @@
 #'   data returned by the API. If TRUE, variables containing absolute paths to
 #'   the preprints web-page ("link_page") and PDF ("link_pdf") are generated
 #'   from the "server", "DOI", and "version" variables returned by the API. The
-#'   "title", "abstract" and "authors" variables are converted to title case.
-#'   Finally, the "type" and "server" variables are dropped.
+#'   The "category", "authors" and "author_corresponding" variables are
+#'   converted to title case. Finally, the "type" and "server" variables are
+#'   dropped.
 #'
 #' @return Dataframe with 1 record per row
 #'
@@ -148,8 +149,9 @@ mx_api_content <- function(from_date = "2013-01-01",
 #'   data returned by the API. If TRUE, variables containing absolute paths to
 #'   the preprints web-page ("link_page") and PDF ("link_pdf") are generated
 #'   from the "server", "DOI", and "version" variables returned by the API. The
-#'   "title", "abstract" and "authors" variables are converted to title case.
-#'   Finally, the "type" and "server" variables are dropped.
+#'   The "category", "authors" and "author_corresponding" variables are
+#'   converted to title case. Finally, the "type" and "server" variables are
+#'   dropped.
 #'
 #' @return Dataframe containing details on the preprint identified by the DOI.
 #'